Add config folder
Includes the pack icon, and changes the name of the program. I attempted change some of the colours on the forge loading screen, but it doesn't seem to work (Likely due to neoforge). This could be fixed with drippy loading screen provided it doesn't do the same thing KJS is trying to do. I could also be completely wrong as to what those config options actually do. At the very least the pack name and image works. A cool loading screen isn't really necessary. If we're going to switch from Neoforge to regular forge, we should do it before phase 1 is released, but the issues we get are fairly minor. We'll just have to check incompatible mods for updates and install them manually.