Quick modlist replacement
Left a few mods I'm experimenting with in there. Requires forge 47.2.6 due to TFC not being compatible with 47.2 and better combat not being compatible with anything lower which means switching from NEOFORGE to FORGE. Technically older mod versions would be fine and we could just wait for neoforge to update, but I can see this becoming a recurring issue in the future that I'm not sure will be fun do deal with. 

Most mods ship with forge in mind first, which isn't usually an issue since both are compatible but when one is made for a version of forge that neoforge hasn't updated to, that's when we start stubbing our toes. I'm not sure how many people care about the drama but since most mods seem to release on forge first, it gives me the idea that neoforge is still an after thought.

Another issue we might run into is mod hosting sites, particularly mods that aren't on forge. When exporting a profile, it warned me that some mods don't originate from curseforge and thus the pack may not be eligible for public distribution. I don't know what mods are in that category (It is likely one of the resource packs) and I don't know if curseforge allows you to upload modpacks with modrinth mods in them. We'll have to cross that bridge when we get to it. While uploading the pack to modrinth won't be a terrible idea when it comes to it, we don't want to exclude people who still prefer to use the curseforge launcher.